jekiiLMS/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1
kombu==5.2.4
numpy==1.24.3
pandas==2.0.1
Pillow==9.4.0
prompt-toolkit==3.0.38
psycopg2==2.9.3
PyMuPDF==1.22.2
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2023.3
redis==4.5.4
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.5.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
